requirements.txt
google-generativeai==0.3.2
IPython
jupyter
langchain-google-genai==0.0.6
pillow
python-dotenv